Move test in the correct lib
